+++ alpharius/data/fmp_client.py
@@ -5,6 +5,7 @@ import threading
 import time
 from typing import Dict, List, Optional
 
+import numpy as np
 import pandas as pd
 import requests
 import retrying
@@ -90,7 +91,8 @@ class FmpClient(DataClient):
                 bars.append(bar)
         bars.sort(key=lambda b: b['date'])
         index = pd.DatetimeIndex([pd.Timestamp(b['date']).tz_localize(TIME_ZONE) for b in bars])
-        data = [[b['open'], b['high'], b['low'], b['close'], b['volume']] for b in bars]
+        data = [[np.float32(b['open']), np.float32(b['high']), np.float32(b['low']), np.float32(b['close']),
+                 np.uint32(b['volume'])] for b in bars]
         return pd.DataFrame(data, index=index, columns=DATA_COLUMNS)
 
     @retrying.retry(stop_max_attempt_number=3,
